backend/src/requirements.txt
Flask
gunicorn
Flask-Cors
py-sudoku